* commit '7df4cbfdff11423ee71fd6e5d529fb68550020a8': (22 commits)
  remove procs
  Good error messages
  More tests, sonar issues
  remove commented code
  remove unused file
  gitignore
  Goal delete fix, production ready docker
  id
  goalid
  Backend changes for E2E tests
  fixed order addProfileQuestionToGoal call inside updateGoal method GoalDAOImpl
  user goal association in db
  Admin goal
  procedure fix
  Changes in testdata, role fixes, more TranslatedNames
  user goal 100% test coverage
  implemented user <-> goal, without unit-test coverage
  procedure fixes
  fixed profile_question relation with goals
  profile questions stored procedures
  ...

+++ API/src/main/java/nl/han/aim/oosevt/lamport/services/goal/GoalServiceImpl.java
@@ -1,10 +1,11 @@
 package nl.han.aim.oosevt.lamport.services.goal;
 
-import nl.han.aim.oosevt.lamport.controllers.goal.dto.CreateGoalRequestDTO;
 import nl.han.aim.oosevt.lamport.controllers.goal.dto.GoalResponseDTO;
 import nl.han.aim.oosevt.lamport.controllers.goal.dto.UpdateGoalRequestDTO;
+import nl.han.aim.oosevt.lamport.controllers.goal.dto.CreateGoalRequestDTO;
 import nl.han.aim.oosevt.lamport.data.dao.goal.GoalDAO;
 import nl.han.aim.oosevt.lamport.data.entity.Goal;
+import nl.han.aim.oosevt.lamport.exceptions.InvalidDTOException;
 import nl.han.aim.oosevt.lamport.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -39,7 +40,7 @@ public class GoalServiceImpl implements GoalService {
     @Override
     public GoalResponseDTO getGoal(int id) {
         final Goal goal = goalDAO.getGoalById(id);
-        if(goal == null) {
+        if (goal == null) {
             throw new NotFoundException();
         }
         return GoalResponseDTO.fromData(goal);
@@ -48,7 +49,7 @@ public class GoalServiceImpl implements GoalService {
     @Override
     public void createGoal(CreateGoalRequestDTO goal) {
         goal.validate();
-        goalDAO.createGoal(goal.getName());
+        goalDAO.createGoal(goal.getName(), goal.getProfileQuestions());
     }
 
     @Override
@@ -56,12 +57,17 @@ public class GoalServiceImpl implements GoalService {
         updateGoalRequestDTO.validate();
         assertGoalExists(updateGoalRequestDTO.getId());
 
-        goalDAO.updateGoal(updateGoalRequestDTO.getId(), updateGoalRequestDTO.getName());
+        goalDAO.updateGoal(updateGoalRequestDTO.getId(), updateGoalRequestDTO.getName(), updateGoalRequestDTO.getProfileQuestions());
     }
 
     @Override
     public void deleteGoal(int id) {
         assertGoalExists(id);
+
+        if(goalDAO.getUserCountByGoalId(id) > 0) {
+            throw new InvalidDTOException();
+        }
+
         goalDAO.deleteGoal(id);
     }
 }
